+++ cosid-core/src/main/java/me/ahoo/cosid/snowflake/SafeJavaScriptSnowflakeId.java
@@ -31,7 +31,7 @@ public final class SafeJavaScriptSnowflakeId {
         return id < JAVA_SCRIPT_MAX_SAFE_NUMBER;
     }
     
-    public static MillisecondSnowflakeId ofMillisecond(long epoch, int timestampBit, int machineBit, int sequenceBit, long machineId, long sequenceResetThreshold) {
+    public static MillisecondSnowflakeId ofMillisecond(long epoch, int timestampBit, int machineBit, int sequenceBit, int machineId, long sequenceResetThreshold) {
         checkTotalBit(timestampBit, machineBit, sequenceBit);
         return new MillisecondSnowflakeId(epoch, timestampBit, machineBit, sequenceBit, machineId, sequenceResetThreshold);
     }
@@ -44,7 +44,7 @@ public final class SafeJavaScriptSnowflakeId {
      * @param machineId 服务实例编号
      * @return MillisecondSnowflakeId
      */
-    public static MillisecondSnowflakeId ofMillisecond(long machineId) {
+    public static MillisecondSnowflakeId ofMillisecond(int machineId) {
         final int timestampBit = MillisecondSnowflakeId.DEFAULT_TIMESTAMP_BIT;
         final int machineBit = MillisecondSnowflakeId.DEFAULT_MACHINE_BIT - 7;
         final int sequenceBit = MillisecondSnowflakeId.DEFAULT_SEQUENCE_BIT - 3;
@@ -52,7 +52,7 @@ public final class SafeJavaScriptSnowflakeId {
         return ofMillisecond(CosId.COSID_EPOCH_SECOND, timestampBit, machineBit, sequenceBit, machineId, SnowflakeId.defaultSequenceResetThreshold(sequenceBit));
     }
     
-    public static SecondSnowflakeId ofSecond(long epoch, int timestampBit, int machineBit, int sequenceBit, long machineId, long sequenceResetThreshold) {
+    public static SecondSnowflakeId ofSecond(long epoch, int timestampBit, int machineBit, int sequenceBit, int machineId, long sequenceResetThreshold) {
         checkTotalBit(timestampBit, machineBit, sequenceBit);
         return new SecondSnowflakeId(epoch, timestampBit, machineBit, sequenceBit, machineId, sequenceResetThreshold);
     }
@@ -65,7 +65,7 @@ public final class SafeJavaScriptSnowflakeId {
      * @param machineId 服务实例编号
      * @return SecondSnowflakeId
      */
-    public static SecondSnowflakeId ofSecond(long machineId) {
+    public static SecondSnowflakeId ofSecond(int machineId) {
         final int timestampBit = SecondSnowflakeId.DEFAULT_TIMESTAMP_BIT;
         final int machineBit = SecondSnowflakeId.DEFAULT_MACHINE_BIT - 7;
         final int sequenceBit = SecondSnowflakeId.DEFAULT_SEQUENCE_BIT - 3;
